requirements.txt
djangorestframework==3.12.2
pydiator-core==1.0.4








